requirements.txt
fastapi
uvicorn
pypdf
paddleocr
paddlepaddle
pytesseract
opencv-python-headless
Pillow
python-multipart
